Library (match.h): recent problems in outer arrangements fixed
Until now, the new k-d tree range search re-structuring of the matching
library were only checked for the inner and full matches, not with the
outer! After doing that test, I noticed that the result of the commands in
the manual is wrong and after some checks, I recognized this is because we
still had one check in the outer arrangements that was inherited from the
past (checking and freeing of 'inrange' within the 'match_kdtree_worker'
function of 'lib/match.c').

With this commit, the problem was solved by keeping the 'inrange' variable
only within the scope of the inner or full arrangements. I also noticed
that an extra check was necessary in the 'ui.c' file of the Match program
and edited (for better clarity) another error message.