NEWS: minor typos and style corrections
Until now, there were a few typos in the NEWS file and the style of the
description of multiple options/function was not consistent: some times the
text started on the same line as the last option/function, and some times
on a new line.

With this commit the typos have been corrected and the style is unified.