Library (kdtree.h): kdtree_range also returns distance
Until now, the recently introduced 'gal_kdtree_range' returned just the
index(es) of the point(s) within the range.

With this commit, it also returns the distance from the desired point.
This helps in situations where a range search is required but also the
information regarding the vicinity is useful. Incidentally, few typos in
the manual have been fixed.