Book: full edit/update of acknowledgment and citation sections
Until now, the acknowledgement section of the book didn't formally mention
the grants that have been very useful for Gnuastro in the last few years
(since I moved to CEFCA), it also confused citations with
acknowledgments. Also, the citations section didn't mention where the full
set of Gnuastro's papers can be found. Finally, a user of Gnuastro does not
need to acknowledge Gnuastro's funders (that is up to Gnuastro itself: the
"acknowledgement" section of the introductoin). A user of Gnuastro should
just cite the papers of the programs they used (as well as the main paper).

With this commit, both sections of the book have gone under a major edit to
address the issues above and the shared text that gets printed with the
'--cite' option in all programs has been updated to remove acknowledgments,
but be more elaborate on how to cite the papers.