Book and PSF scripts: acknowledgements and citations updated
Until now, the PSF scripts only recommended citing Infante-Sainz+2020, but
with the publication of Eskandarlou+2025, it was also necessary to mention
that in the citations and the book along with the former. In particular,
all the new lessons learnt and the many figures are very helpful for a new
user. On the other hand, the acknowledgemnt section that was added in the
last commit needed some further corrections and clarifications.

With this commit, all the issues above have been addressed: the PSF scripts
now contain the Eskandarlou+2025 paper both within the script and in the
book. Also, the the acknowledgements section is now more complete after
sharing with the CEFCA administration.