Configuration: development mode for next release
Until now (in the last commit), Gnuastro 0.24 was released. To have a
smooth start in new developments, the library SONAME version needs to be
updated as well the comments on the book's PDF first page.

With this commit, both have been implemented along with some edits in the
release-checklist for the next releases.