Book: using Unicode character for Angstroms
Until now, the book used the LaTeX command '\AA' for angstroms. However,
after converting to HTML and uploading the pages with the 0.24 release, I
noticed that for some reason '\AA' is no longer being interpretted properly
and is mistakenly written raw! I also noticed that due to some apparent
changes in the GNU HTML styles, the "Download" title of the top Gnuastro
page was not at the start of the line due to the image above it.

With this commit, the Unicode 'Å' is used instead to avoid problems. This
was the best solution I could find in the limited available time and works
on the web browsers and the command-line. The fix was manually inserted in
the relevant HTML pages and uploaded independently so viewers do not
problems until the next Gnuastro release. In the top Gnuastro page an empty
line was also added to fix the "Download" title problem.