Book: definition of ST magnitudes also added after AB magnitudes
Until now, in the "brightness, flux and magnitude" section of MakeCatalog
(that defines the core units used in astronomy), we only talked about the
AB magnitdue system. But some people also use the ST magnitude and without
it, that section was not complete.

With this commit, a new part has been added to also define and distingush
the AB from ST magnitudes.