astscript-fits-view: support globbing in ds9region
Until now, a syntax like '--ds9region=a-*.reg' was expanded as one region
and several arguments (that DS9 would interpret as FITS images). Therefore,
only one region would load and DS9 would complain about not being able to
open the rest. This was because the assumed delimiter between the region
files was a comma (,). Using (double) quotes didn't solve the issue.

With this commit, '--ds9region="a-*.reg"' is correcly expanded as several
regions. To enable this it was necessary to add support for a
SPACE-separated list and the documentation has been updated with examples
of usage.