ConvertType: SLS colormap variation for not having white
Until now, the only variation of the SLS color map was its inverse. While
the SLS color map has many advantages, one issue it has is that its maximum
value is white (which is also the default color to display NaNs). Because
the default background of printable papers is white, we want the NaNs to
display as white, but not the maximum values. As a result, with the SLS
color, the maximum values are displayed as white.

With this commit, two new color maps have been added to ConvertType to fix
this problem: 'sls-no-white' and 'sls-inverse-no-white'. In the process, I
also found some other issues that have also been fixed in this commit.

  - The description of '--cmappgfplots' did not have a minimal set of steps
    to load the produced colormap into PGFPlots.

  - In the example script of section "Full script of annotations on figure"
    there was a problem with loading the colormap if it had a dash.

  - There were some TABs at the start of some example code lines (that do
    not display similarly on differnet viewers).