Library (wcsdistortion.h): solve memory bug
Until now, running 'astfits img.fits --wcsdistortion=SIP' on an image with
an odd number of pixels on at least one of the two axis caused a crash with
a "double free or corruption (out)" error.

With this commit, the issue is solved by accounting the extra element that
is necessary for odd-numbered images. It was actually a
'heap-buffer-overflow' (as detected by '-fsanitize=address').
Incidentally, a memory leak in the same file is also removed (it was found
while trying to debug the crash).

This fixes bug #67776.