Library (arithmetic.h): all rotate-coord operands can be constant
Until now, when calling the 'rotate-coord' operator, it was not possible to
keep all the coordiantes constant and only change the rotation angle. This
was because we would always write the output in the first two
arguments. This was done to save memory, but it would make the usage very
complicated (especially with the given example where only the angle
changed).

With this commit, the issue was solved by allocating a dedicated output
array. The usage example in the book has also been corrected and elaborated
to help in the understanding/usage.