pointing-simulate: polishing code and minor fix in book
Until now, in the book the description of the '--hook-warp-after' option of
this script wrongly mentioned that the output should be called
'$TOWARP'. Also, the actual script didn't account for sub-directories in
the temporary directory, and the comment lines in the corresponding
Makefile were longer than our coding convention requirements.

With this commit, all three issues have been fixed.