Book: acknowledgement and citation info added in copyright page
Until now, these were explained in detail within the book, but such legal
things are best mentioned at the very start (in the copyright page).

With this commit, a short explanation has been added there and for official
releases, the DOI of the book will also be printed on the title page
(currently commented). Furthermore, since ADS is migrating to SciX, the
links to the full list of Gnuastro papers in the book and output of
'--cite' now points to SciX.